pyproject.toml
[project]
name = "flashcard-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "google-genai>=1.33.0",
    "jinja2>=3.1.6",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
